Change return type of sampgdk_get_natives() from 'size_t' to 'int'

+++ include/sampgdk/core.h
@@ -20,8 +20,6 @@
 #include <sampgdk/amx.h>
 #include <sampgdk/export.h>
 
-#include <stddef.h>
-
 #define AMX_EXEC_GDK (-10)
 
 /* These should be called only once, e.g. in Load() and Unload() respectively. */
@@ -41,7 +39,7 @@ SAMPGDK_EXPORT void SAMPGDK_CALL sampgdk_process_timers();
 SAMPGDK_EXPORT void SAMPGDK_CALL sampgdk_process_plugin_timers(void *plugin);
 
 SAMPGDK_EXPORT const AMX_NATIVE_INFO *SAMPGDK_CALL sampgdk_get_natives();
-SAMPGDK_EXPORT size_t SAMPGDK_CALL sampgdk_num_natives();
+SAMPGDK_EXPORT int SAMPGDK_CALL sampgdk_num_natives();
 
 typedef void (*sampgdk_logprintf_t)(const char *format, ...);
 SAMPGDK_EXPORT sampgdk_logprintf_t sampgdk_logprintf;
